requirements.txt
python-dotenv
ollama
sentence-transformers
lancedb
pyarrow
pypdf
python-docx
loguru
tqdm
openpyxl